requirements.txt
pyserial>=3.5
picamera2
requests
python-dotenv
opencv-python
